requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Bcrypt==0.7.1
gotrue==2.6.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
postgrest==0.16.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
realtime==1.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.12.2
websockets==12.0
Werkzeug==3.0.3
WTForms==3.1.2
